requirements.txt
absl-py==2.3.1
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
APScheduler==3.11.2
asgiref==3.9.2
astunparse==1.6.3
bcrypt==5.0.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
bottle==0.13.4
bs4==0.0.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
clr_loader==0.2.8
colorama==0.4.6
cryptography==46.0.3
cssselect==1.3.0
distro==1.9.0
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
fake-useragent==2.2.0
fastapi==0.121.1
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
fsspec==2025.10.0
gast==0.6.0
google-pasta==0.2.0
graphviz==0.21
greenlet==3.3.0
groq==1.0.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
keras==3.11.3
libclang==18.1.1
lxml==6.0.2
lxml_html_clean==0.4.3
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
mysql-connector-python==9.5.0
namex==0.1.0
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.3.4
openai==2.7.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
parse==1.20.2
pefile==2024.8.26
pillow==11.3.0
portalocker==3.2.0
protobuf==6.33.0
proxy_tools==0.1.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydotplus==2.0.2
pyee==11.1.1
Pygments==2.19.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
PyMySQL==1.1.2
pyparsing==3.2.4
pyppeteer==2.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pythonnet==3.0.5
pywebview==6.1
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
requests-html==0.10.0
rich==14.2.0
safetensors==0.6.2
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
sqlparse==0.5.3
starlette==0.49.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
textblob==0.19.0
tf_keras==2.20.1
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
unicorn==2.1.4
urllib3==1.26.20
uv==0.8.20
uvicorn==0.38.0
vaderSentiment==3.3.2
w3lib==2.3.1
webencodings==0.5.1
websockets==10.4
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
zipp==3.23.0
